refactor: separate wallet_ffi cucumber steps into own file
